Resume Analyzer/Resume Analyzer/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.2
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docx2txt==0.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
nbclient==0.6.5
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.23.0
packaging==21.3
pafy==0.5.5
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathy==0.6.2
pdfminer.six==20220524
pdfminer3==2018.12.3.0
pdfplumber==0.7.1
phonenumbers==8.12.51
pickleshare==0.7.5
Pillow==9.1.1
plotly==5.9.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyresparser==1.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.1
resume-parser==0.8.4
rich==12.4.4
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.3.0
stemming==1.0.1
streamlit==1.10.0
streamlit-tags==1.2.8
tenacity==8.0.1
terminado==0.15.0
thinc==8.0.17
threadpoolctl==3.1.0
tika==1.24
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
Wand==0.6.7
wasabi==0.9.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
youtube-dl==2021.12.17
zipp==3.8.0
